Fix dirent.h clang-devel build failures
Originally, fdopendir was declared in dirent.h
using __DARWIN_ALIAS_I. I think this was needed
because the original implementation had local
variables of type struct statbuf, and so its
translation unit needed to be compiled multiple
times, once for each inode size, and each version
would be differently labelled, and __DARWIN_ALIAS_I
was probably needed to select the correctly
labelled version of fdopendir. But I am guessing.

The current implementation no longer uses any
struct statbuf local variables, and so is only
compiled once. I think this means that it is no
longer necessary to use __DARWIN_ALIAS_I when
declaring fdopendir. Perhaps surprisingly, it
still worked with __DARWIN_ALIAS_I in place.
So maybe I'm guessing wrong.

However, the clang-devel port failed to compile on
macOS-10.9, complaining about the attempt to
overload a function (fdopendir) that differed only
in return type (i.e., LegacySupport's wrapper DIR
struct, as opposed to the system DIR struct). And
maybe I've misunderstood the error message. And
this compiler error didn't appear on other legacy
platforms. So the problem isn't definitely understood.

I might be wrong about why and when __DARWIN_ALIAS_I
is needed, but it seemed that it possibly isn't needed
anymore (if it ever was).

This patch stops declaring fdopendir via __DARWIN_ALIAS_I
and it doesn't break LegacySupport on macOS-10.6 at least.
It compiles and the tests pass. If it works on other legacy
platforms, this should fix clang-devel's build failure by
eliminating the attempt to overload fdopendir.

The redefinition of __MPLS_INODE64 to nothing when __DARWIN_INODE64
is not defined is to match macOS-10.4.